chore: remove console.log

+++ contracts/8.7/xALPACA.sol
@@ -102,7 +102,7 @@ contract xALPACA is Initializable, ReentrancyGuardUpgradeable, OwnableUpgradeabl
 
     pointHistory.push(Point({ bias: 0, slope: 0, timestamp: block.timestamp, blockNumber: block.number }));
 
-    uint256 _decimals = IBEP20(_token).decimals();
+    uint8 _decimals = IBEP20(_token).decimals();
     require(_decimals <= 255, "bad decimals");
     decimals = _decimals;
